move to irdl.core_attrs
